refactor(ATT-434): split connectionConfigurationScreen god file into focused units
Split the ~700-line connectionConfigurationScreen.cpp into cohesive
translation units, keeping construction/lifecycle in the main file:
  _wifi.cpp      WiFi scan + network dropdown + loop()
  _keyboard.cpp  on-screen keyboard + text-area focus
  _save.cpp      save widgets, validation, save flow
  _pinlock.cpp   PIN-lock overlay + callbacks

Pure code move, behavior preserved. All 4 board envs build green.
Bump firmware version 1.3.13 -> 1.3.14 for OTA.
