+++ apps/attractap/firmware/src/display/screens/connectionConfiguration/connectionConfigurationScreen.cpp
@@ -1,11 +1,11 @@
 #include "connectionConfigurationScreen.hpp"
-#include <cstring>
-#include "../../../network/wifi/wifi.hpp"
 
-static const char *WIFI_DROPDOWN_LOADING = "Suche WLANs...";
-static const char *WIFI_DROPDOWN_EMPTY = "Keine Netzwerke gefunden";
-static const char *WIFI_DROPDOWN_SCAN_FAILED = "WLAN Scan fehlgeschlagen";
-static const uint32_t WIFI_SCAN_TIMEOUT_MS = 10000;
+// Screen construction (tabs/widgets) and lifecycle. Behaviour-specific logic
+// lives in sibling translation units:
+//   _wifi.cpp      WiFi scan + network dropdown
+//   _keyboard.cpp  on-screen keyboard + text-area focus
+//   _save.cpp      save widgets, validation, save flow
+//   _pinlock.cpp   PIN-lock overlay + callbacks
 
 void ConnectionConfigurationScreen::init()
 {
@@ -48,7 +48,7 @@ void ConnectionConfigurationScreen::init()
    lv_obj_set_style_text_color(labelForWifiSelectNetwork, lv_color_white(), LV_PART_MAIN | LV_STATE_DEFAULT);
 
    this->wifiSelectNetwork = lv_dropdown_create(wifiTab);
-   lv_dropdown_set_options(this->wifiSelectNetwork, WIFI_DROPDOWN_LOADING);
+   lv_dropdown_set_options(this->wifiSelectNetwork, "Suche WLANs...");
    lv_obj_set_width(this->wifiSelectNetwork, lv_pct(100));
    lv_obj_set_height(this->wifiSelectNetwork, LV_SIZE_CONTENT);
    lv_obj_set_align(this->wifiSelectNetwork, LV_ALIGN_CENTER);
@@ -157,9 +157,6 @@ void ConnectionConfigurationScreen::init()
    lv_obj_t *containerForSaveButton = this->createSaveContainer(apiTab);
    this->createSaveButton(containerForSaveButton);
 
-   // Also add a save button to the API tab header/footer container for consistency
-   // (re-use existing containerForSaveButton)
-
    // Device tab
    lv_obj_t *deviceTab = lv_tabview_add_tab(this->tabs, "Geraet");
    lv_obj_set_flex_flow(deviceTab, LV_FLEX_FLOW_COLUMN);
@@ -231,434 +228,9 @@ void ConnectionConfigurationScreen::init()
    }
 }
 
-
 lv_obj_t *ConnectionConfigurationScreen::getScreen()
 {
    return this->screen;
-}
-
-void ConnectionConfigurationScreen::loop()
-{
-   if (!this->wifiSelectNetwork || !this->wifiScanRequested || this->wifiScanCompleted)
-   {
-      return;
-   }
-
-   if (Wifi::isScanning())
-   {
-      if (this->wifiScanStartMs > 0 && (millis() - this->wifiScanStartMs) > WIFI_SCAN_TIMEOUT_MS)
-      {
-         this->wifiScanCompleted = true;
-         this->wifiScanRequested = false;
-         this->wifiDropdownHasNetworks = false;
-         lv_dropdown_set_options(this->wifiSelectNetwork, WIFI_DROPDOWN_SCAN_FAILED);
-      }
-      return;
-   }
-
-   this->wifiScanCompleted = true;
-   this->wifiScanRequested = false;
-   this->populateWifiDropdown();
-}
-
-void ConnectionConfigurationScreen::startWifiScan()
-{
-   if (!this->wifiSelectNetwork)
-   {
-      return;
-   }
-
-   this->wifiScanRequested = true;
-   this->wifiScanCompleted = false;
-   this->wifiScanStartMs = millis();
-   this->wifiDropdownHasNetworks = false;
-   lv_dropdown_set_options(this->wifiSelectNetwork, WIFI_DROPDOWN_LOADING);
-   Wifi::startScan();
-}
-
-void ConnectionConfigurationScreen::populateWifiDropdown()
-{
-   if (!this->wifiSelectNetwork)
-   {
-      return;
-   }
-
-   Wifi::WifiScanResult scan = Wifi::getKnownWifiNetworks();
-   String options = "";
-   String savedSSID = Settings::getNetworkConfig().ssid;
-   uint8_t selectedIndex = 0;
-   bool selectedFound = false;
-
-   String uniqueSsids[Wifi::MAX_KNOWN_WIFI_NETWORKS];
-   uint8_t uniqueCount = 0;
-
-   for (uint8_t i = 0; i < scan.count; i++)
-   {
-      const String &ssid = scan.networks[i].ssid;
-      if (ssid.length() == 0)
-      {
-         continue;
-      }
-
-      bool isDuplicate = false;
-      for (uint8_t j = 0; j < uniqueCount; j++)
-      {
-         if (uniqueSsids[j] == ssid)
-         {
-            isDuplicate = true;
-            break;
-         }
-      }
-      if (isDuplicate)
-      {
-         continue;
-      }
-
-      if (uniqueCount >= Wifi::MAX_KNOWN_WIFI_NETWORKS)
-      {
-         break;
-      }
-
-      uniqueSsids[uniqueCount] = ssid;
-      uniqueCount++;
-
-      if (options.length() > 0)
-      {
-         options += "\n";
-      }
-      options += ssid;
-
-      if (!selectedFound && savedSSID.length() > 0 && ssid == savedSSID)
-      {
-         selectedIndex = uniqueCount - 1;
-         selectedFound = true;
-      }
-   }
-
-   if (options.length() == 0)
-   {
-      this->wifiDropdownHasNetworks = false;
-      lv_dropdown_set_options(this->wifiSelectNetwork, WIFI_DROPDOWN_EMPTY);
-      return;
-   }
-
-   this->wifiDropdownHasNetworks = true;
-   lv_dropdown_set_options(this->wifiSelectNetwork, options.c_str());
-   if (selectedFound)
-   {
-      lv_dropdown_set_selected(this->wifiSelectNetwork, selectedIndex);
-   }
-}
-
-void ConnectionConfigurationScreen::onTextAreaEvent(lv_event_t *e)
-{
-   ConnectionConfigurationScreen *self = static_cast<ConnectionConfigurationScreen *>(lv_event_get_user_data(e));
-   if (!self)
-      return;
-   lv_event_code_t code = lv_event_get_code(e);
-   lv_obj_t *target = (lv_obj_t *)lv_event_get_target(e);
-
-   if (code == LV_EVENT_FOCUSED || code == LV_EVENT_CLICKED)
-   {
-      self->showKeyboardFor(target);
-   }
-}
-
-void ConnectionConfigurationScreen::onWifiDropdownEvent(lv_event_t *e)
-{
-   ConnectionConfigurationScreen *self = static_cast<ConnectionConfigurationScreen *>(lv_event_get_user_data(e));
-   if (!self || !self->wifiSelectNetwork || !self->wifiSSID)
-   {
-      return;
-   }
-
-   lv_event_code_t code = lv_event_get_code(e);
-   if (code != LV_EVENT_VALUE_CHANGED)
-   {
-      return;
-   }
-
-   if (!self->wifiDropdownHasNetworks)
-   {
-      return;
-   }
-
-   char selected[64] = {0};
-   lv_dropdown_get_selected_str(self->wifiSelectNetwork, selected, sizeof(selected));
-   if (selected[0] == '\0')
-   {
-      return;
-   }
-
-   lv_textarea_set_text(self->wifiSSID, selected);
-}
-
-void ConnectionConfigurationScreen::onKeyboardEvent(lv_event_t *e)
-{
-   ConnectionConfigurationScreen *self = static_cast<ConnectionConfigurationScreen *>(lv_event_get_user_data(e));
-   if (!self)
-      return;
-
-   lv_event_code_t code = lv_event_get_code(e);
-   if (code == LV_EVENT_READY || code == LV_EVENT_CANCEL)
-   {
-      self->hideKeyboardIfNoFocus();
-   }
-   else if (code == LV_EVENT_VALUE_CHANGED)
-   {
-      lv_obj_t *kb = (lv_obj_t *)lv_event_get_target(e);
-      if (!kb)
-         return;
-
-      lv_keyboard_mode_t mode = lv_keyboard_get_mode(kb);
-      if (mode == LV_KEYBOARD_MODE_TEXT_UPPER)
-      {
-         uint32_t btn_id = lv_keyboard_get_selected_button(kb);
-         if (btn_id != LV_BUTTONMATRIX_BUTTON_NONE)
-         {
-            const char *btn_txt = lv_keyboard_get_button_text(kb, btn_id);
-            if (btn_txt && btn_txt[0] >= 'A' && btn_txt[0] <= 'Z' && btn_txt[1] == '\0')
-            {
-               lv_keyboard_set_mode(kb, LV_KEYBOARD_MODE_TEXT_LOWER);
-            }
-         }
-      }
-   }
-}
-
-// removed continue button handler (replaced by Save buttons on each tab)
-
-static bool isEmpty(const char *text)
-{
-   return !(text && text[0] != '\0');
-}
-
-static bool hostnameLooksValid(const char *text)
-{
-   if (!text || text[0] == '\0')
-      return false;
-   return true;
-}
-
-static bool pinLooksValid(const char *text)
-{
-   if (!text)
-      return false;
-   size_t len = strlen(text);
-   if (len < 4)
-      return false;
-   for (size_t i = 0; i < len; ++i)
-   {
-      if (text[i] < '0' || text[i] > '9')
-         return false;
-   }
-   return true;
-}
-
-void ConnectionConfigurationScreen::onSaveButtonEvent(lv_event_t *e)
-{
-   ConnectionConfigurationScreen *self = static_cast<ConnectionConfigurationScreen *>(lv_event_get_user_data(e));
-   if (!self)
-      return;
-
-   lv_event_code_t code = lv_event_get_code(e);
-   if (code != LV_EVENT_CLICKED)
-      return;
-
-   const char *ssidText = lv_textarea_get_text(self->wifiSSID);
-   const char *passwordText = lv_textarea_get_text(self->wifiPassword);
-   const char *hostText = lv_textarea_get_text(self->serverHostname);
-   const char *devicePinText = self->devicePin ? lv_textarea_get_text(self->devicePin) : "";
-
-   String hostValue = String(hostText ? hostText : "");
-   if (hostValue.startsWith("https://"))
-   {
-      hostValue.remove(0, 8);
-      lv_textarea_set_text(self->serverHostname, hostValue.c_str());
-   }
-
-   if (hostValue.startsWith("http://"))
-   {
-      hostValue.remove(0, 7);
-      lv_textarea_set_text(self->serverHostname, hostValue.c_str());
-   }
-
-   bool hostValid = !hostValue.isEmpty() && hostnameLooksValid(hostValue.c_str());
-   bool devicePinValid = pinLooksValid(devicePinText);
-
-   // Update label colors
-   if (self->labelForServerHostname)
-   {
-      lv_obj_set_style_text_color(self->labelForServerHostname,
-                                  hostValid ? self->labelForServerHostnameDefaultColor : lv_color_hex(0xFF0000),
-                                  LV_PART_MAIN | LV_STATE_DEFAULT);
-   }
-   if (self->labelForDevicePin)
-   {
-      lv_obj_set_style_text_color(self->labelForDevicePin,
-                                  devicePinValid ? self->labelForDevicePinDefaultColor : lv_color_hex(0xFF0000),
-                                  LV_PART_MAIN | LV_STATE_DEFAULT);
-   }
-
-   // Focus first invalid input and ensure visible
-   if (!hostValid)
-   {
-      // Switch to API tab, focus server hostname
-      if (self->tabs)
-         lv_tabview_set_act(self->tabs, 1, LV_ANIM_ON);
-      lv_keyboard_set_textarea(self->keyboard, self->serverHostname);
-      self->showKeyboardFor(self->serverHostname);
-      lv_obj_scroll_to_view_recursive(self->serverHostname, LV_ANIM_ON);
-      return;
-   }
-   if (!devicePinValid)
-   {
-      if (self->tabs)
-         lv_tabview_set_act(self->tabs, 2, LV_ANIM_ON);
-      lv_keyboard_set_textarea(self->keyboard, self->devicePin);
-      self->showKeyboardFor(self->devicePin);
-      lv_obj_scroll_to_view_recursive(self->devicePin, LV_ANIM_ON);
-      return;
-   }
-
-   // All valid -> call callback if provided
-   if (self->onSaveCallback)
-   {
-      ConnectionConfigurationScreen::ConnectionConfig cfg;
-      cfg.ssid = String(ssidText);
-      cfg.password = String(passwordText);
-      cfg.host = hostValue;
-      cfg.useSSL = lv_obj_has_state(self->useSSLSwitch, LV_STATE_CHECKED);
-      cfg.devicePin = String(devicePinText);
-      cfg.beeperEnabled = lv_obj_has_state(self->beeperEnabled, LV_STATE_CHECKED);
-      self->onSaveCallback(cfg);
-   }
-}
-
-void ConnectionConfigurationScreen::showKeyboardFor(lv_obj_t *targetTextArea)
-{
-   if (!targetTextArea)
-      return;
-
-   // Retarget keyboard
-   lv_keyboard_set_textarea(this->keyboard, targetTextArea);
-
-   // Choose mode based on field
-   if (targetTextArea == this->wifiSSID)
-   {
-      lv_keyboard_set_mode(this->keyboard, LV_KEYBOARD_MODE_TEXT_LOWER);
-   }
-   else if (targetTextArea == this->wifiPassword)
-   {
-      lv_keyboard_set_mode(this->keyboard, LV_KEYBOARD_MODE_TEXT_LOWER);
-   }
-   else if (targetTextArea == this->serverHostname)
-   {
-      lv_keyboard_set_mode(this->keyboard, LV_KEYBOARD_MODE_TEXT_LOWER);
-   }
-   else if (targetTextArea == this->devicePin)
-   {
-      lv_keyboard_set_mode(this->keyboard, LV_KEYBOARD_MODE_NUMBER);
-   }
-
-   // Show keyboard and shrink tabs
-   lv_obj_clear_flag(this->keyboard, LV_OBJ_FLAG_HIDDEN);
-   if (this->tabs)
-   {
-      lv_obj_set_height(this->tabs, lv_pct(50));
-   }
-
-   // Ensure focused input is visible after layout changes
-   lv_obj_update_layout(this->screen);
-   lv_obj_scroll_to_view_recursive(targetTextArea, LV_ANIM_ON);
-}
-
-void ConnectionConfigurationScreen::hideKeyboardIfNoFocus()
-{
-   lv_obj_add_flag(this->keyboard, LV_OBJ_FLAG_HIDDEN);
-   if (this->tabs)
-   {
-      lv_obj_set_height(this->tabs, lv_pct(100));
-   }
-}
-
-lv_obj_t *ConnectionConfigurationScreen::createSaveButton(lv_obj_t *parent)
-{
-   lv_obj_t *save = lv_button_create(parent);
-   lv_obj_set_width(save, 100);
-   lv_obj_set_height(save, 50);
-   lv_obj_set_align(save, LV_ALIGN_CENTER);
-   lv_obj_add_flag(save, LV_OBJ_FLAG_SCROLL_ON_FOCUS);
-   lv_obj_remove_flag(save, LV_OBJ_FLAG_SCROLLABLE);
-   lv_obj_set_style_bg_color(save, lv_color_hex(0x00FF00), LV_PART_MAIN | LV_STATE_DEFAULT);
-   lv_obj_set_style_bg_opa(save, 255, LV_PART_MAIN | LV_STATE_DEFAULT);
-   lv_obj_add_event_cb(save, &ConnectionConfigurationScreen::onSaveButtonEvent, LV_EVENT_CLICKED, this);
-
-   lv_obj_t *label = lv_label_create(save);
-   lv_obj_set_width(label, LV_SIZE_CONTENT);
-   lv_obj_set_height(label, LV_SIZE_CONTENT);
-   lv_obj_set_align(label, LV_ALIGN_CENTER);
-   lv_label_set_text(label, "Speichern");
-   lv_obj_set_style_text_color(label, lv_color_hex(0x000000), LV_PART_MAIN | LV_STATE_DEFAULT);
-   lv_obj_set_style_text_opa(label, 255, LV_PART_MAIN | LV_STATE_DEFAULT);
-
-   return save;
-}
-
-lv_obj_t *ConnectionConfigurationScreen::createSaveContainer(lv_obj_t *parent)
-{
-   lv_obj_t *container = lv_obj_create(parent);
-   lv_obj_remove_style_all(container);
-   lv_obj_set_width(container, lv_pct(100));
-   lv_obj_set_height(container, LV_SIZE_CONTENT);
-   lv_obj_set_align(container, LV_ALIGN_CENTER);
-   lv_obj_set_flex_flow(container, LV_FLEX_FLOW_ROW);
-   lv_obj_set_flex_align(container, LV_FLEX_ALIGN_END, LV_FLEX_ALIGN_START, LV_FLEX_ALIGN_START);
-   lv_obj_remove_flag(container, LV_OBJ_FLAG_CLICKABLE);
-   lv_obj_remove_flag(container, LV_OBJ_FLAG_SCROLLABLE);
-   return container;
-}
-
-void ConnectionConfigurationScreen::setOnSaveCallback(std::function<void(const ConnectionConfigurationScreen::ConnectionConfig &)> onSaveCallback)
-{
-   this->onSaveCallback = onSaveCallback;
-}
-
-void ConnectionConfigurationScreen::disablePinLock()
-{
-   this->pinLockEnabled = false;
-   if (!this->pinLockOverlay)
-      return;
-   lv_obj_add_flag(this->pinLockOverlay, LV_OBJ_FLAG_HIDDEN);
-}
-
-void ConnectionConfigurationScreen::enablePinLock()
-{
-   this->pinLockEnabled = true;
-   if (!this->pinLockOverlay)
-      return;
-   lv_obj_clear_flag(this->pinLockOverlay, LV_OBJ_FLAG_HIDDEN);
-}
-
-bool ConnectionConfigurationScreen::onPinLockConfirmCallback(String pin)
-{
-   String devicePin = Settings::getDeviceConfig().passCode;
-   bool matches = pin == devicePin;
-
-   if (!matches)
-   {
-      // delay to slow down brute force attacks
-      delay(5000);
-      return false;
-   }
-
-   this->disablePinLock();
-   return true;
-}
-
-void ConnectionConfigurationScreen::setOnCancelPinLockCallback(std::function<void()> onCancelPinLockCallback)
-{
-   this->onCancelPinLockCallback = onCancelPinLockCallback;
 }
 
 String ConnectionConfigurationScreen::getName()
